Create Nwokeoma-Praise-Uchechukwu-2024714143
This Python program implements a grading system using “if” and “else” statements. It takes a student's score as input and print the corresponding grades based on the following scale:

- 70 - 100: Grade A
- 60 - 69: Grade B
- 50 - 59: Grade C
- 45 - 49: Grade D
- 40 - 44: Grade E
- 20 - 39: Grade F

For each valid score, the program also prints “YES” to confirm the grading was successful. If the score is outside the valid range, it prints “Invaid score” and “NO